requirements.txt
antspyx==0.4.2
blinker==1.4
certifi==2024.2.2
charset-normalizer==3.3.2
chart-studio==1.1.0
#command-not-found==0.3
contourpy==1.2.1
cryptography==3.4.8
cycler==0.12.1
#dbus-python==1.2.18
dill==0.3.8
distro==1.7.0
#distro-info==1.1+ubuntu0.2
docopt==0.6.2
fonttools==4.51.0
httplib2==0.20.2
idna==3.6
imageio==2.34.0
importlib-metadata==4.6.4
jeepney==0.7.1
joblib==1.3.2
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
#lazy_loader==0.4
matplotlib==3.8.4
more-itertools==8.10.0
#netifaces==0.11.0
networkx==3.2.1
nibabel==5.2.1
numpy==1.26.4
oauthlib==3.2.0
packaging==24.0
pandas==2.2.1
patsy==0.5.6
pillow==10.3.0
plotly==5.20.0
#PyGObject==3.42.1
PyJWT==2.3.0
pykwalify==1.8.0
pyparsing==2.4.7
# pyradiomics==3.0.1 - failing here, manually installed - 8N
#python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
pytz==2024.1
PyWavelets==1.5.0
#PyYAML==5.4.1
requests==2.31.0
retrying==1.3.4
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.13.0
SecretStorage==3.3.1
SimpleITK==2.3.1
six==1.16.0
statsmodels==0.14.1
#systemd-python==234
tenacity==8.2.3
threadpoolctl==3.4.0
tifffile==2024.2.12
tzdata==2024.1
#ubuntu-pro-client==8001
#ufw==0.36.1
#unattended-upgrades==0.1
urllib3==2.2.1
wadllib==1.3.6
webcolors==1.13
zipp==1.0.0
